Chatbot/requirements.txt
bs4
flask
requests
line-bot-sdk
requests_html



